split ui function
